Added DM swing support
