--- requirements.txt
@@ -1,8 +1,0 @@
-configparser~=5.0.2
-discord~=1.7.3
-dhooks~=1.1.4
-requests~=2.27.1
-Pillow~=8.3.2
-mysql-connector-python~=8.0.27
-praw~=7.5.0
-asyncpraw~=7.2.0
