+++ src/assets.py
@@ -158,3 +158,5 @@ main_role_ids = {
 }
 
 fcb_team_ids = ['PUR', 'ORD', 'MSU', 'MCH', 'HRW', 'PAN']
+
+valid_positions = ['P', 'C', '1B', '2B', '3B', 'SS', 'LF', 'CF', 'RF', 'DH', 'PH', 'PR']
